Creacao da blueprint db
